GEODE-6588: Mass cleanup of static analyzer warnings.

+++ geode-core/src/distributedTest/java/org/apache/geode/cache/management/MemoryThresholdsDUnitTest.java
@@ -2076,7 +2076,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
               public String load(LoaderHelper<Integer, String> helper) throws CacheLoaderException {
                 Integer expectedInvocations = (Integer) helper.getArgument();
                 final int actualInvocations = numLoaderInvocations.getAndIncrement();
-                if (expectedInvocations.intValue() != actualInvocations) {
+                if (expectedInvocations != actualInvocations) {
                   throw new CacheLoaderException("Expected " + expectedInvocations
                       + " invocations, actual is " + actualInvocations);
                 }
@@ -2153,7 +2153,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
             ((InternalResourceManager) getCache().getResourceManager()).getHeapMonitor();
         assertFalse(hmm.getState().isCritical());
 
-        int expectedInvocations = expected.intValue();
+        int expectedInvocations = expected;
         {
           Integer k = new Integer(4);
           assertEquals(k.toString(), r.get(k, new Integer(expectedInvocations++)));
@@ -2269,7 +2269,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
             return expectedInvocations1;
           }
         });
-    expectedInvocations.set(ex.intValue());
+    expectedInvocations.set(ex);
 
     ex = (Integer) ds1
         .invoke(new SerializableCallable("Invoke loader from datastore, non-critical") {
@@ -2289,7 +2289,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
             return expectedInvocations2;
           }
         });
-    expectedInvocations.set(ex.intValue());
+    expectedInvocations.set(ex);
 
     accessor.invoke(addExpectedException);
     ds1.invoke(addExpectedException);
@@ -2323,7 +2323,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
             return expectedInvocations5;
           }
         });
-    expectedInvocations.set(ex.intValue());
+    expectedInvocations.set(ex);
 
     ex = (Integer) accessor.invoke(new SerializableCallable(
         "During critical state on datastore, assert accesor load behavior") {
@@ -2340,7 +2340,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
         return expectedInvocations7;
       }
     });
-    expectedInvocations.set(ex.intValue());
+    expectedInvocations.set(ex);
 
     ex = (Integer) ds1.invoke(
         new SerializableCallable("Set safe state on datastore, assert local load behavior") {
@@ -2365,7 +2365,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
             return expectedInvocations8;
           }
         });
-    expectedInvocations.set(ex.intValue());
+    expectedInvocations.set(ex);
 
     accessor.invoke(new SerializableCallable(
         "Data store in safe state, assert load behavior, accessor sets critical state, assert load behavior") {
@@ -2444,7 +2444,7 @@ public class MemoryThresholdsDUnitTest extends ClientServerTestCase {
             public String load(LoaderHelper<Integer, String> helper) throws CacheLoaderException {
               Integer expectedInvocations = (Integer) helper.getArgument();
               final int actualInvocations = numLoaderInvocations.getAndIncrement();
-              if (expectedInvocations.intValue() != actualInvocations) {
+              if (expectedInvocations != actualInvocations) {
                 throw new CacheLoaderException("Expected " + expectedInvocations
                     + " invocations, actual is " + actualInvocations);
               }
