+++ geode-core/src/distributedTest/java/org/apache/geode/cache/query/dunit/QueryAPITestPartitionResolver.java
@@ -30,7 +30,7 @@ public class QueryAPITestPartitionResolver implements PartitionResolver {
 
   @Override
   public Serializable getRoutingObject(EntryOperation opDetails) {
-    return (((Integer) opDetails.getKey()).intValue()
+    return ((Integer) opDetails.getKey()
         % QueryUsingFunctionContextDUnitTest.numOfBuckets);
   }
